+++ samples/src/main/kotlin/jp/shiguredo/sora/sample/facade/SoraAudioChannel.kt
@@ -3,6 +3,7 @@ package jp.shiguredo.sora.sample.facade
 import android.content.Context
 import android.os.Handler
 import jp.shiguredo.sora.sample.option.SoraRoleType
+import jp.shiguredo.sora.sdk.channel.SoraCloseEvent
 import jp.shiguredo.sora.sdk.channel.SoraMediaChannel
 import jp.shiguredo.sora.sdk.channel.data.ChannelAttendeesCount
 import jp.shiguredo.sora.sdk.channel.option.SoraAudioOption
@@ -45,8 +46,8 @@ class SoraAudioChannel(
             handler.post { listener?.onConnect(this@SoraAudioChannel) }
         }
 
-        override fun onClose(mediaChannel: SoraMediaChannel) {
-            SoraLogger.d(TAG, "[audio_channel] @onClose")
+        override fun onClose(mediaChannel: SoraMediaChannel, closeEvent: SoraCloseEvent) {
+            SoraLogger.d(TAG, "[audio_channel] @onClose $closeEvent")
             disconnect()
         }